backend/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
bcrypt
pyotp
cryptography
mysql-connector-python
psycopg2-binary
